my-ai-project/requirements.txt
openai
langchain
autogen
python-dotenv
